feat: 인증 허용 및 타입 변경

Closes S12P21A604-222

See merge request s12-fintech-finance-sub1/S12P21A604!242

+++ rabbit-server/src/main/java/com/rabbit/coin/domain/dto/request/CoinPermitRequestDTO.java
@@ -3,6 +3,8 @@ package com.rabbit.coin.domain.dto.request;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.*;
 
+import java.math.BigInteger;
+
 @Schema(description = "코인 permit 요청 DTO")
 @Getter
 @Setter
@@ -12,7 +14,7 @@ import lombok.*;
 public class CoinPermitRequestDTO {
     private String owner;
     private String spender;
-    private long value;
-    private long deadline;
+    private BigInteger value;
+    private BigInteger deadline;
     private String signature;
 }
